sched/fix: Combine the two if conditions into one in energy_diff
The two if conditions do exactly the same and the first one contains a continue.
So if src_cpu and dst_cpu belong to the same sd_ea sched group, only one if
condition is carried out any way. If they belong to different sd_ea sched groups
the if condition is carried out twice.

The test 'eenv->{src,dst}_cpu != -1' is not necessary without energy_diff being
used in the load balancer.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>