sched/fix: Create eas specific cpu_in_sg test function
To be easier on the reviewer's eyes, create a test function to check
if a cpu belongs to a sched group. The test of the cpu against -1 is
only necessary if it is used in load balancer context.

cpumask_test_cpu does not deal with 'cpu = -1' but if
CONFIG_DEBUG_PER_CPU_MAPS is enabled it would print out a warning.

The function cannot be used troughout the scheduler because 'cpu = -1'
is no valid value in these cases.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>