sched/cpufreq_sched: fix passive updates, add to trace event
Passive updates (capacity changes which do not request an immediate
frequency change) are currently broken since they do not get
propagated to the total capacity vote value for the CPU. They are
only set in the sched class' vote value.

Fix this by re-calculating the total capacity vote for passive
updates as well in update_cpu_capacity_request. This has the side
benefit of capturing passive changes with the trace event there
as well, although a new value is needed in it to show whether the
change is passive or not.

Signed-off-by: Steve Muckle <steve.muckle@linaro.org>