sched/cpufreq_sched: fix stuck vote on passive capacity changes
When update_cpu_capacity_request is called with request = 0 the
associated frequency change is not made. But the updated capacity
vote for the cpu should still be recorded. Fix this.

Also the pcpu_capacity percpu variable is superfluous now that
there is a percpu structure containing votes from each sched class
and a total.

Signed-off-by: Steve Muckle <smuckle@linaro.org>