sched/cpufreq_sched: always update frequency when capacity changes
The cpufreq_sched_set_cap() function currently bails if the CPU
specified is not the CPU with the maximum capacity request in the
policy. This can lead to the following situation (assume CPU0 and
CPU1 are part of the same policy):

 - CPU0 gets heavily loaded, requests cap = 1024 (fmax)
 - CPU1 gets lightly loaded, requests cap = 10
 - CPU0's load goes away, requests cap = 0
 - CPU1's load of 10 persists for a long time

In this sequence we will run at fmax unnecessarily until CPU1's
load eventually changes or CPU0's load rises to 10 or above.

Signed-off-by: Steve Muckle <smuckle@linaro.org>