sched/fair: update scale invariance of pelt
The current implementation of load tracking invariance scales the load
tracking value with current frequency and uarch performance (only for
utilization) of the local CPU.

One main result of the current formula is that the figures are capped by the
current capacity of the local CPU. This limitation is the main reason of not
including the uarch invariance in the calculation of load_avg.

Instead of scaling the complete value of PELT algo, we should only scale the
running time by current freq and uarch of the local CPU. It seems more correct
to only scale the running time because the non running time (sleeping or
waiting for a runqueue) of a task is the same whatever the current freq and the
compute capacity of the local CPU.

Then, one main advantage of this change is that the load of a task can reach
max value whatever the current freq and the uarch of the local CPU. It will
just take more time at a lower freq than the max freq or on a "little" CPU
compared to a "big" one. The load and the utilization stay invariant across
system but with a wider range of value.

With this change, we don't have to test if a CPU is overloaded or not in order
to use one metric or another as all metrics are always valid.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>