BACKPORT: cpufreq: governor: initialize/destroy timer_mutex with 'shared'
timer_mutex is required to be initialized only while memory for 'shared'
is allocated and in a similar way it is required to be destroyed only
when memory for 'shared' is freed.

There is no need to do the same every time we start/stop the governor.
Move code to initialize/destroy timer_mutex to the relevant places.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
( backported from 5e4500d8dba16d88b528cf037566b84747ec23f0 )
Signed-off-by: Chris Redpath <chris.redpath@arm.com>

Change-Id: I9ec30996b05152dc6d5dd1198be2b073c19d9213