firmware: arm_scpi_test: Stress tests for SCP and SCPI driver
These can be run by passing kernel commandline parameters, e.g.

  arm_scpi_test.stress_time=60 arm_scpi_test.run=3

or from a shell after boot like:

  echo 60 > /sys/module/arm_scpi_test/parameters/stress_time
  echo 3 > /sys/module/arm_scpi_test/parameters/run

The 'run' parameter gives the number of the test case to run, or -1
to indicate all tests. Setting run to zero will stop currently running
tests.

'stress_time' is the number of seconds each stress test is run for
and if omitted a default for each test will be used.

Test progress is output with pr_info, pr_err etc, so will appear in the
kernel log and on a console if the kernel log is also directed there.

Signed-off-by: Jon Medhurst <tixy@linaro.org>