DEBUG: FIX: sched/fair: report only root RQ's CPU util
The RQ utilization is reported for all RQs, thus when TGs are in use we
get events reporting values for different sub-RQs without the chance to
distinguish them when it comes to plot these signals.

This patch enforce to report always (and only) the root RQ utilization.
Indeed, this is the most useful signal for most of our plots.

TODO: report the TG name in the sched_load_avg_cpu event.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>