WIP: UTIL_EST: sched/fair: add support to return the estimated CPU utilization
The estimated RQ utilization can be used for some customers which are
interested into an estimated signal which is more table than PELT.

This patch adds a new getter method cpu_util_est(cpu) which is defined
to always return the MAX between the CPU estimated utilization and the
current value of util_avg. This new signal is usable both when tasks are
RUNNABLE as well as when the RQ is empty.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>