sched: Introduce SD_ASYM_CAPACITY sched_domain topology flag
Add a helper flag to the sched_domain hierarchy indicating which domains
that span cpus of different capacity (e.g. big.LITTLE). This information
is currently only available through interating through the capacities of
all the cpus.

This flag is not guaranteed to be set bottom. It is more likely not to
as it shouldn't be set for domains spanning cpu of identical capacities.