sched/cpufreq_sched: add throttle tunable, default throttling to 0
The cpufreq framework already ensures that frequency transition
requests are serialized, so there is no need to try and serialize them
by delaying them according to the advertised transition
latency. However, it may still be useful to throttle frequency
transitions for power reasons, so add a sysfs tunable for this
purpose.

Signed-off-by: Steve Muckle <smuckle@linaro.org>