sched/cpufreq_sched: support mixed fast/slow path
Even when the cpufreq driver is fast enough to be called in the
scheduler hot path it may sometimes be forced to the slow path
due to throttling or the cpufreq driver being busy with another
transition.

Add support for throttling in the fast path as well as concurrent
requests from both the fast and slow paths.

Signed-off-by: Steve Muckle <smuckle@linaro.org>