WIP: kernel/sched/fair: add hysteresis filter to task_fits
When the task utilisation is in the range of a CPU capacity margin, we can
experiments oscillations which produces noisy migrations between clusters.
Such migrations could be particularly dangerous for performances since
we invalidate local caches.

This patch introduces an pair of margins which are used to create an hysteresis
filter around the threshold margin. The standard 20% threshold is enlarged
by a +-2% which seems to be quite good to make more stable the allocation
of a ramp task.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>