sched/{fair,tune}: move schedtune_nornalize_energy into fair.c
The energy normalization function is required to get the proper values
for the P-E space filtering function to work. That normalization is part
of the hot wakeup path and currently implemented with a function call.

Moving the normalization function into fair.c allows the compiler to
inline that code thus reducing overheads in the wakeup hot path.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>