TBT: fix idle state estimation for destination CPU/Cluster
It's not easy to estimate the idle state once a task as been moved away from
a CPU. However, for sure we know that, at least the destination CPU and
its higher level SGs will be in the shallowest idle state.

This patch ensure that the idle state of a destination CPU and cluster
is always properly estimated. To force idle_idx=0 we exploit the code
already in place for the computation of the capacity variations.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>