BACKPORT: cpufreq: intel_pstate: Replace timers with utilization update callbacks
Instead of using a per-CPU deferrable timer for utilization sampling
and P-states adjustments, register a utilization update callback that
will be invoked from the scheduler on utilization changes.

The sampling rate is still the same as what was used for the deferrable
timers, so the functional impact of this patch should not be significant.

Based on an earlier patch from Srinivas Pandruvada.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
( backported from a4675fbc4a7abe072ac6ba38c252f22a91ebcd94 )
Signed-off-by: Chris Redpath <chris.redpath@arm.com>

Change-Id: Ibafb5c1d8e0efc17d45b45fb92e104069c5d0875