sched/fix: Change energy model data check function
EAS requests that all cpus in a sched group have the same energy model
data. Since all cpus have to use the same sge, only the sg->sge pointers
have to be compared to make sure they are identical. So get rid of the
comparison of the individual idle and capacity state vector elements.

This patch has to be squashed into "sched: Initialize energy data
structures".

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>