sched/deadline: move task bw on fallback and wu migrations
Steve's test showed that we still have holes in the ac_bw tracking.
What can happen is that, while a DEADLINE task is sleeping its
cpus_allowed mask change (e.g., via cpusets) and at wakeup we call
select_fallback_rq() to decide where to wake it up. Since we didn't
remove the task bw from the rq it was one before sleeping, and we don't
also add it to the fallback rq, the ac_bw tracking is broken. Same
happens when we migrate the task in select_task_rq_dl().

This patch fixes it by keeping track of the select_fallback_rq()
decision (with a new task_struct field) and uses this information in
migrate_task_rq_dl() (also added) to move the task bw between the old rq
and the fallback one. For the wakeup migration case we simply use the
target cpu to do the same.

Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>
Reported-by: Wanpeng Li <wanpeng.li@hotmail.com>
Reported-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Juri Lelli <juri.lelli@arm.com>