WIP: UTIL_EST: sched/fair: add support enable/disable estimation capping
The util_avg signal for a CPU can sometimes overpass the SCHED_LOAD_SCALE
value, for example when a task with a certain utilization is enqueued in
a RQ which has a spare capacity lower than the task utilization.

This patch allows to defined if the RQ util_est value should be kept
within the [0..SCHED_LOAD_SCALE] range, thus being eventually capped.

Capping the util_est value can result in an underestimation of the utilization.
This happens all the time we dequeue a task which was enqueued with a capped
utilization value.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>