BACKPORT: cpufreq: governor: Avoid atomic operations in hot paths
Rework the handling of work items by dbs_update_util_handler() and
dbs_work_handler() so the former (which is executed in scheduler
paths) only uses atomic operations when absolutely necessary.  That
is, when the policy is shared and dbs_update_util_handler() has
already decided that this is the time to queue up a work item.

In particular, this avoids the atomic ops entirely on platforms where
policy objects are never shared.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
( backported from e4db2813d2e558b6b6bee464308678a57732b390 )
Signed-off-by: Chris Redpath <chris.redpath@arm.com>

Change-Id: Ib3ebc73340688949a41c92c14afd04aba7a8a3a7