BACKPORT: cpufreq: governor: Fix nice contribution computation in dbs_check_cpu()
The contribution of the CPU nice time to the idle time in dbs_check_cpu()
is computed in a bogus way, as the code may subtract current and previous
nice values for different CPUs.

That doesn't matter for cases when cpufreq policies are not shared,
but may lead to problems otherwise.

Fix the computation and simplify it to avoid taking unnecessary steps.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
( backported from 679b8fe43a6b723787cae1d9599ed776d7ce238b )
Signed-off-by: Chris Redpath <chris.redpath@arm.com>

Change-Id: Iff5695b9d19f7125f315b8caf6d589c9366d44d7