BACKPORT: cpufreq: governor: Move io_is_busy to struct dbs_data
The io_is_busy governor tunable is only used by the ondemand governor
and is located in the ondemand-specific data structure, but it is
looked at by the common governor code that has to do ugly things to
get to that value, so move it to struct dbs_data and modify ondemand
accordingly.

Since the conservative governor never touches that field, it will
be always 0 for that governor and it won't have any effect on the
results of computations in that case.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
( backported from 8847e038c1d19c20dda0d7a590e31ffa528da8a5 )
Signed-off-by: Chris Redpath <chris.redpath@arm.com>

Change-Id: Ib840cacbceea086085883dc7385ef82e4ec07e41