BACKPORT: cpufreq: governor: Drop unnecessary checks from show() and store()
The show() and store() routines in the cpufreq-governor core don't need
to check if the struct governor_attr they want to use really provides
the callbacks they need as expected (if that's not the case, it means a
bug in the code anyway), so change them to avoid doing that.

Also change the error value to -EBUSY, if the governor is getting
removed and we aren't allowed to store any more changes.

Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
( backported from f737236b128cac7c355d0650a98c42ae4313f3f1 )
Signed-off-by: Chris Redpath <chris.redpath@arm.com>

Change-Id: I5b25d4357e761286980f002f99057e9e39c81e15