sched/fix: Make sure that a single group sd does not disturb scheduling
There are to occurrences where the code makes assumptions that sd->parent
is used for scheduling operations. Since this is not true for a single
sched group (sg) sched domain (sd) which sole purpose is to hold energy
data, test if the parent sd exists and has more than one sg.

For EAS in particular this allows that the system-wide over-utilization
indicator gets set to false on systems with such a single sg sd as the
root sd (e.g. SYS sd on Hi6220).

This patch has to be squashed into "sched/debug: EAS & cpu hotplug
interoperability".

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>