sched/fix: Add a BUG_ON to check idle_idx
Since we're relying on cpu idle framework information to figure out in
which idle state we're currently in add a check that the idle state, when
used as an index to the energy model idle state vector, is inside the
boundaries of this vector.

We don't have this issue with the capacity state vector since we're only
use EAS functionality to operate on them.

This patch has to be squashed into "sched: Determine the current
sched_group idle-state".

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>