sched/fair: Clean up the logic in fix_small_imbalance()
Avoid the need to add scaled_busy_load_per_task on both sides of the if
condition to determine whether imbalance has to be set to
busiest->load_per_task or not.

The imbn variable was introduced with commit 2dd73a4f09be ("[PATCH]
sched: implement smpnice") and the original if condition was

  if (max_load - this_load >= busiest_load_per_task * imbn)

which morved into the current version where scaled_busy_load_per_task
is to be found on both sides of the if condition.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>