sched/fair: Skip first sched group with target in select_idle_sibling()
The sched domain (sd) to start the 'for each lower sd' loop is the last
level cache domain from the viewpoint of the target cpu. The sd setup
makes sure that the viewpoint cpu is always member of the first sched
group (sg).

So skip the first sg before iterating over all sg's of the sd instead
of figuring that the target cpu is member of sg while iterating over
all sg's cpus.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>