TBT: EXPERIMENTAL: EAS: max energy increase tunable
This is an ugly patch for a proof-of-concept experiment on accepting
scheduling decisions even when we have an energy increase, provided
that the evaluated energy increase is below a specified threshold.

A new sysfs tunable is exposed via:
   /proc/sys/kernel/sched_cfs_max_nrgi
which allows to specify the _absolute_ energy variation which is to
be accepted.

NOTE: both positive and negative thresholds can be specified, where:
- positive: allows to accept sched candidate which increase the
            system energy, which could be beneficial to exit local
            minimums
- negative: allows to reject sched candidate which decrease a small
            amount of system energy, which could be beneficial to
            avoid migration penalties when the chances to save energy
            are really limited.

Caveats: the current tunable uses absolute energy diff numbers, thus
the values to specify depends on the specific target and its energy
model.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>