schedtune: handle accounting error in schedtune boost groups
This works around the case where a task is enqueued while a member
of one sched-tune group, but the task is migrated to another group
before being dequeued. This would manifest as an attempt to
decrement below zero which would be ignored -- but the accounting
for the old boost group would never be decremented resulting in
CPU boosts that would never return to zero (possibly causing
higher frequencies).

A proper fix is coming with a callback that adjusts the accounting
when the task is moved to a different group.

Change-Id: I18be893c4433d7fd3b3b23da7c734f37acda3c43