sched/fair: Add energy_diff dead-zone margin
It is not worth the overhead to migrate tasks for tiny insignificant
energy savings. To prevent this, an energy margin is introduced in
energy_diff() which effectively adds a dead-zone that rounds tiny energy
differences to zero. Since no scale is enforced for energy model data
the margin can't be absolute. Instead it is defined as n% energy saving
compared to the current total estimated energy consumption. n is 5 by
default (so 5%), but it can be changed in
/proc/sys/kernel/sched_nrg_dead_zone .