BACKPORT: cpufreq: governor: CPUFREQ_GOV_STOP never fails
None of the cpufreq governors currently in the tree will ever fail
an invocation of the ->governor() callback with the event argument
equal to CPUFREQ_GOV_STOP (unless invoked with incorrect arguments
which doesn't matter anyway) and it is rather difficult to imagine
a valid reason for such a failure.

Accordingly, rearrange the code in the core to make it clear that
this call never fails.

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
( backported from 45482c703b8c7e86e339e92f186cadd9a735fd0e )
Signed-off-by: Chris Redpath <chris.redpath@arm.com>

Change-Id: Icf4f70178ca8512f675c6731fc579ecd14a04b6a