gem5,ARM: Enable the CPUFreq framework and corresponding drivers
In this changeset we enable the CPUFreq framework, as well as the
corresponding gem5 drivers, in the gem5_defconfig for arm. We enable
the following goverors:

* Performance (Default)
* Powersave
* Userspace
* Ondemand
* Conservative

The default governor is Performance as this will not adjust the
frequency, and will leave it at the maximum frequency. Therefore,
unless the governor is changed, future simulations should run as
before.

Change-Id: I02f2bac7e38824a9b4273dce58c1fda512e15176
Signed-off-by: Sascha Bischoff <sascha.bischoff@arm.com>