sched: trace: Replace load_avg with walt util in task trace
Use the existing load_avg member to display the walt utilisation as seen
when walt stats are used to make scheduling decisions.

walt doesn't really maintain a live average, rather it's calculated when
necessary. The calculation here is done slightly differently so we can
get a result in 32 bits.

This is just done like this so we can avoid changing upper tools and still
see the differences between walt signals and pelt utilisation signals.