SchedTune: add sentinels
This patch adds some members in the task_struct to keep track of
enqueue/dequeue events for each task. At each enqueue we keep track of the
CPU/boostgroup we added +1 and verify at dequeue time that we do a -1 on the
same CPU/boostgroup. These are the invariant which should ensure a proper
accounting.

In schedtune_can_attach a RUNNABLE task is chancing its boost group while
running on the same CPU where it has been enqueue previously. Thus sentinels
are verified and the boostgroup index updated.

Each time sentinels does not match the invariant we report a WARNing event
on the trace... of course these events should never happen... and that's
the case for me at least on JUNO, but still I see a drift in the accounting.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>