SchedTune: FIX accounting for throttled entities
If an entity is enqueued into a throttled TG we do not account for it being
RUNNABLE. Technically a throttled task is not RUNNABLE, but accounting for
runnable-throttled tasks can be quite complicated and deserves a better
analysis to understand if the increased complexity is compensated by
reasonable benefits.

For the time being we account only for RUNNABLE tasks independently from
their throttled state, which provides a completely working solution at least
for systems not using the CFS bandwidth controller (e.g. Android).

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>