SchedTune: FIX exit/migration race condition
Apparently, a task can be force migrated while it's executing a do_exit()
syscall. In this cases it could happen that the task is flagged PF_EXITING
before a dequeue, which disable our accounting for the task migrating its
CPU.

This is a kind-of-NON-patch which seems to make things working... let's keep
track of this status while we improve the checks on task status at dequeue
and enqueue time.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>